+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Constructors of Props and Widgets are now public
 
+### Fixed
+- Vertical/Horizontal stretch now respect spaces of sibling and stretch as long as siblings have space
+
 ## [0.0.14]
 
 ### Added